Do not -lpng
This project uses SDL_Image to load images.

libpng doesn't need to and should not be linked directly in this case, as this causes issues if the libpng version on the device is not the same as in the toolchain. E.g. gcw0 toolchain has libpng14 but newer RG350 firmwares will have the latest libpng.